Refactor user-client API call and add MyPosts and MyPets pages with error handling and loading states
