+++ app/(pages)/(content)/profile/my-pets/[id]/page.tsx
@@ -4,34 +4,31 @@
 import Banners from '@/components/banners';
 import { useEffect, useState } from 'react';
 
-import { getPosts } from '@/utils/posts.http';
 import { Pet } from '@/types/pet';
-import { useRouter } from 'next/navigation';
-import { useAuth } from '@/contexts/authContext';
+import { useParams, useRouter } from 'next/navigation';
 import Loading from '@/app/loading';
 import PetCard from '@/components/petCard/pet-card';
 import LabeledSelect from '@/components/labeled-selected';
-import Footer from '@/components/footer';
-import { Post } from '@/types/post';
+import { getPetsByUserId } from '@/utils/pets.http';
+import { error } from 'console';
 
 
 
 const fetchContentData = async (
-    setPosts: React.Dispatch<React.SetStateAction<Post[]>>,
+    setPets: React.Dispatch<React.SetStateAction<Pet[]>>,
     setLoading: React.Dispatch<React.SetStateAction<boolean>>,
-    setPostsError: React.Dispatch<React.SetStateAction<string | null>>,
+    setPetsError: React.Dispatch<React.SetStateAction<string | null>>,
     userId: string,
 ) => {
 
 
     try {
         // Cargar posts del usuario
-        const postParams = { user: userId }; // Usamos el ID del usuario actual
-        const postData = await getPosts(postParams);
-        setPosts(Array.isArray(postData) ? postData : []);
+        const petData = await getPetsByUserId(userId);
+        setPets(Array.isArray(petData) ? petData : []);
     } catch (err) {
         console.error("Error al cargar posts:", err);
-        setPostsError("No se pudieron cargar las publicaciones."); // 👈 Manejo de error separado
+        setPetsError("No se pudieron cargar las publicaciones."); // 👈 Manejo de error separado
     } finally {
         setLoading(false);
     }
@@ -41,38 +38,29 @@ export default function MyPostsPage() {
     const mascotas = ["Todos", "Conejo", "Perro", "Gato"];
     const edades = ["0-1 años", "1-3 años", "3-6 años", "6+ años"];
 
-    const { authToken, user, loading: authLoading } = useAuth();
-    const [posts, setPosts] = useState<Post[]>([]);
+    const {id} = useParams();
+
+    const [pets, setPets] = useState<Pet[]>([]);
     const [loading, setLoading] = useState<boolean>(true);
-    const router = useRouter();
-    const [postsError, setPostsError] = useState<string | null>(null);
+    const [petsError, setPetsError] = useState<string | null>(null);
 
     const [selectedCiudad, setSelectedCiudad] = useState<string | null>(null);
     const [selectedMascota, setSelectedMascota] = useState<string | null>(null);
     const [selectedEdad, setSelectedEdad] = useState<string | null>(null);
 
-    useEffect(() => {
-        if (!authLoading && !authToken) {
-            console.log("authLoading", authLoading);
-            console.log("authToken", authToken);
-            router.push("/auth/login");
-        }
 
-    }, [authToken, authLoading, router]);
 
 
     useEffect(() => {
-        if (authLoading || !authToken || !user?.id) return;
-        console.log("authLoading", authLoading);
-        fetchContentData(setPosts, setLoading, setPostsError, user.id);
-    }, [authToken, authLoading, user?.id]);
+        if(!id) return
+        fetchContentData(setPets, setLoading, setPetsError, id.toString());
+    }, []);
 
-    if (authLoading) {
+    if (loading) {
         return Loading();
     }
-
-    if (!user) return;
     const bannerImages = ["/banner1.png", "/banner2.png", "/banner3.png", "/banner4.png"]
+    if(loading) return <Loading />
 
     return (
         <div className='flex flex-col gap-5'>
@@ -111,9 +99,10 @@ export default function MyPostsPage() {
 
             <section>
                 <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-5 gap-12 px-12 py-4">
-                    {/* 🔥 Mapeo de las mascotas */}
-                    {posts.length > 0 ? (
-                        posts.map((post) => (
+                    
+                    {petsError? <p className="text-center col-span-full">Hubo un error al cargar las mascotas</p> : 
+                    pets.length > 0 ? (
+                        pets.map((post) => (
                             <PetCard key={post.id} post={post} />
                         ))
                     ) : (
