+++ app/(pages)/(content)/profile/my-posts/[id]/page.tsx
@@ -5,33 +5,30 @@ import Banners from '@/components/banners';
 import { useEffect, useState } from 'react';
 
 import { getPosts } from '@/utils/posts.http';
-import { Pet } from '@/types/pet';
-import { useRouter } from 'next/navigation';
-import { useAuth } from '@/contexts/authContext';
-import Loading from '@/app/loading';
+import { useParams, useRouter } from 'next/navigation';
 import PetCard from '@/components/petCard/pet-card';
 import LabeledSelect from '@/components/labeled-selected';
-import Footer from '@/components/footer';
 import { Post } from '@/types/post';
-import { getPetsByUserId } from '@/utils/pets.http';
+import Loading from '@/app/loading';
 
 
 
 const fetchContentData = async (
-    setPets: React.Dispatch<React.SetStateAction<Pet[]>>,
+    setPosts: React.Dispatch<React.SetStateAction<Post[]>>,
     setLoading: React.Dispatch<React.SetStateAction<boolean>>,
-    setPetsError: React.Dispatch<React.SetStateAction<string | null>>,
+    setPostsError: React.Dispatch<React.SetStateAction<string | null>>,
     userId: string,
 ) => {
 
 
     try {
         // Cargar posts del usuario
-        const petData = await getPetsByUserId(userId);
-        setPets(Array.isArray(petData) ? petData : []);
+        const postParams = { user: userId }; // Usamos el ID del usuario actual
+        const postData = await getPosts(postParams);
+        setPosts(Array.isArray(postData) ? postData : []);
     } catch (err) {
         console.error("Error al cargar posts:", err);
-        setPetsError("No se pudieron cargar las publicaciones."); // 👈 Manejo de error separado
+        setPostsError("No se pudieron cargar las publicaciones."); // 👈 Manejo de error separado
     } finally {
         setLoading(false);
     }
@@ -40,38 +37,24 @@ export default function MyPostsPage() {
     const ciudades = ["Encarnación", "Asunción", "Luque", "Fernando Zona Sur"];
     const mascotas = ["Todos", "Conejo", "Perro", "Gato"];
     const edades = ["0-1 años", "1-3 años", "3-6 años", "6+ años"];
+    const {id} = useParams();
 
-    const { authToken, user, loading: authLoading } = useAuth();
-    const [pets, setPets] = useState<Pet[]>([]);
+    const [posts, setPosts] = useState<Post[]>([]);
     const [loading, setLoading] = useState<boolean>(true);
-    const router = useRouter();
-    const [petsError, setPetsError] = useState<string | null>(null);
+    const [postsError, setPostsError] = useState<string | null>(null);
 
     const [selectedCiudad, setSelectedCiudad] = useState<string | null>(null);
     const [selectedMascota, setSelectedMascota] = useState<string | null>(null);
     const [selectedEdad, setSelectedEdad] = useState<string | null>(null);
 
-    useEffect(() => {
-        if (!authLoading && !authToken) {
-            console.log("authLoading", authLoading);
-            console.log("authToken", authToken);
-            router.push("/auth/login");
-        }
 
-    }, [authToken, authLoading, router]);
 
 
     useEffect(() => {
-        if (authLoading || !authToken || !user?.id) return;
-        console.log("authLoading", authLoading);
-        fetchContentData(setPets, setLoading, setPetsError, user.id);
-    }, [authToken, authLoading, user?.id]);
-
-    if (authLoading) {
-        return Loading();
-    }
-
-    if (!user) return;
+        if(!id) return
+        fetchContentData(setPosts, setLoading, setPostsError, id.toString());
+    }, []);
+    if(loading) return <Loading />
     const bannerImages = ["/banner1.png", "/banner2.png", "/banner3.png", "/banner4.png"]
 
     return (
@@ -111,13 +94,13 @@ export default function MyPostsPage() {
 
             <section>
                 <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-5 gap-12 px-12 py-4">
-                    {/* 🔥 Mapeo de las mascotas */}
-                    {pets.length > 0 ? (
-                        pets.map((post) => (
+                {postsError? <p className="text-center col-span-full">Hubo un error al cargar las publicaciones</p> : 
+                    posts.length > 0 ? (
+                        posts.map((post) => (
                             <PetCard key={post.id} post={post} />
                         ))
                     ) : (
-                        <p className="text-center col-span-full">Cargando mascotas...</p>
+                        <p className="text-center col-span-full">Cargando las publicaciones...</p>
                     )}
 
                 </div>
